Replace ratio with root
Your videos make it seem as if you do not think too highly of the root test, not the ratio test.  Original: It is no secret that I do not think too highly of the ratio test.  Proposed:  It is no secret that I do not think too highly of the root test.